Make ModelObserver single-threaded.
This CL puts the burden of thread management on
the caller of ModelObserver instead of handling it
in the ModelObserver itself. ThreadLocal is very
expensive from a performance perspective and
we can avoid ThreadLocal calls on every
observeReads() call by pushing it to when threads
are created. Each thread will have to use its
own ModelObserver.

Was able to remove now-unused API temporaryReadObserver().

ModelObserver also changed its commit callback
mechanism to use something like an Executor.

Fixes: 146144712
Test: ScrollerBenchmark#first_layout, ModelObserverBenchmark

Change-Id: Ica33d559a037dec3c14a7c388a6b335020108874